Add files via upload
Multiple Input Multiple Output (MIMO) radar employs multiple antennas in both transmission and reception. Employing the virtual array, MIMO radar can have a great degree of freedom. In this homework, a MIMO radar is designed to be able to estimate the Direction of Arrivals (DoA) of a specific target in the field of view (FoV). Two methods could be used to estimate the DoV: FFT based method, and Backprojection method.